+++ Yuki Theme.Core/Database/DatabaseManager.cs
@@ -46,7 +46,6 @@ namespace Yuki_Theme.Core.Database
 				SetValueToDatabase (ke, Settings.ASKTOSAVE, "true");
 				SetValueToDatabase (ke, Settings.SAVEASOLD, "true");
 				SetValueToDatabase (ke, Settings.SHOWPREVIEW, "true");
-				SetValueToDatabase (ke, Settings.PREVIEWIMAGES, "false");
 			}
 		}
 
@@ -90,7 +89,6 @@ namespace Yuki_Theme.Core.Database
 			AddToDictionary (ref dictionary, key, Settings.ASKTOSAVE, "true");
 			AddToDictionary (ref dictionary, key, Settings.SAVEASOLD, "true");
 			AddToDictionary (ref dictionary, key, Settings.SHOWPREVIEW, "true");
-			AddToDictionary (ref dictionary, key, Settings.PREVIEWIMAGES, "false");
 
 			return dictionary;
 		}