+++ Yuki Theme.Core/Settings.cs
@@ -33,7 +33,6 @@ public static class Settings
 	public static bool         askToSave;
 	public static bool         saveAsOld;
 	public static bool         showPreview;
-	public static bool         previewImages;
 	
 	#endregion
 	
@@ -75,14 +74,9 @@ public static class Settings
 	public const int SAVEASOLD = 29;
 	public const int SHOWPREVIEW = 30;
 	
-	/// <summary>
-	/// Previewing theme with wallpaper and sticker can slow down the IDE. So, you can disable showing wallpaper and sticker on preview.  
-	/// </summary>
-	public const int PREVIEWIMAGES = 31;
-	
 	
 	public const  double current_version     = 7.0;
-	public const  string current_version_add = "beta";
+	public const  string current_version_add = "";
 	public static string next_version        = "";
 
 	#endregion
@@ -151,7 +145,6 @@ public static class Settings
 		askToSave = bool.Parse (data [ASKTOSAVE]);
 		saveAsOld = bool.Parse (data [SAVEASOLD]);
 		showPreview = bool.Parse (data [SHOWPREVIEW]);
-		previewImages = bool.Parse (data [PREVIEWIMAGES]);
 
 		CLI.selectedItem = data [ACTIVE];
 		var os = 0;
@@ -193,7 +186,6 @@ public static class Settings
 		dict.Add (ASKTOSAVE, askToSave.ToString ());
 		dict.Add (SAVEASOLD, saveAsOld.ToString ());
 		dict.Add (SHOWPREVIEW, showPreview.ToString ());
-		dict.Add (PREVIEWIMAGES, previewImages.ToString ());
 		database.UpdateData (dict);
 		if (CLI_Actions.onBGIMAGEChange != null) CLI_Actions.onBGIMAGEChange ();
 		if (CLI_Actions.onSTICKERChange != null) CLI_Actions.onSTICKERChange ();